Support the VK_EXT_hdr_metadata extension.
This extension requires macOS 10.15. For some reason, even though
Apple said that the new Apple TV supports HDR10 and DolbyVision, the
`CAEDRMetadata` type underlying this is not supported on either iOS or
tvOS.

Also set the `wantsExtendedDynamicRangeContent` property on macOS to
enable HDR output. It's not only set if HDR metadata is attached, but
also if an HDR color space is selected.