Rename most MVKPixelFormats member functions to eliminate naming redundancy.
Many of the names of the MVKPixelFormats member functions where inherited from
function names in mvk_datatypes.h, which is a C API, therefore must have unique
function names, and uses parameter types as part of the names to distinguish.

Since MVKPixelFormats is a C++ class, we can do away with that and make use
of function overloading to simplify the function names.

I forgot to to this when creating MVKPixelFormats in the first place.