Restrict MTLTextureUsagePixelFormatView to only being disabled if the device supports native swizzles.
Also, if the device supports native swizzles, always use the swizzle rather than a format reinterpretation.  This part of the change will impact how swizzles are handled regardless of how allowImageCopyReinterpretation is set on devices that support native swizzle.  I can't think of a reason not to do that and it's necessary to make this change viable.