MVKPixelFormats: Correct format features for GBGR/BGRG formats.
Some things I missed in my review.

These formats use implicit reconstruction, and we don't support explicit
reconstruction here yet. If there's a demand for it, I or someone else
can implement it, but I don't expect it to be an issue.

Because it uses implicit reconstruction, chroma sampling is determined
by the implementation, i.e. Metal. My testing shows that Metal on AMD
and Intel uses midpoint reconstruction. Based on this, I think NV will
be the same.

Finally, according to the Vulkan spec:

> Implicit reconstruction takes place by the samples being interpolated,
> as required by the filter settings of the sampler, *except that
> `chromaFilter` takes precedence for the chroma samples*. [emphasis
> added]

Since Metal obviously uses the same filters for chroma and luma, we
can't support the separate reconstruction filter feature here.